[chore #159875590] rename package to lowercase
