--- .idea/inspectionProfiles/Project_Default.xml
@@ -1,26 +1,0 @@
-<component name="InspectionProjectProfileManager">
-  <profile version="1.0">
-    <option name="myName" value="Project Default" />
-    <inspection_tool class="PyPackageRequirementsInspection" enabled="true" level="WARNING" enabled_by_default="true">
-      <option name="ignoredPackages">
-        <value>
-          <list size="13">
-            <item index="0" class="java.lang.String" itemvalue="coverage" />
-            <item index="1" class="java.lang.String" itemvalue="nose" />
-            <item index="2" class="java.lang.String" itemvalue="aniso8601" />
-            <item index="3" class="java.lang.String" itemvalue="atomicwrites" />
-            <item index="4" class="java.lang.String" itemvalue="more-itertools" />
-            <item index="5" class="java.lang.String" itemvalue="pluggy" />
-            <item index="6" class="java.lang.String" itemvalue="py" />
-            <item index="7" class="java.lang.String" itemvalue="gunicorn" />
-            <item index="8" class="java.lang.String" itemvalue="attrs" />
-            <item index="9" class="java.lang.String" itemvalue="coveralls" />
-            <item index="10" class="java.lang.String" itemvalue="flask-restplus" />
-            <item index="11" class="java.lang.String" itemvalue="docopt" />
-            <item index="12" class="java.lang.String" itemvalue="certifi" />
-          </list>
-        </value>
-      </option>
-    </inspection_tool>
-  </profile>
-</component>
